User and UserRole: @Valid added, @Data deleted, dead code deleted, UserRole id type changed

+++ src/main/java/com/eni/bookhub/repository/UserRoleRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface UserRoleRepository extends JpaRepository<UserRole, Long> {
+public interface UserRoleRepository extends JpaRepository<UserRole, Integer> {
     public Optional<UserRole> getUserRoleByRoleName(String role);
-    public Optional<UserRole>  getUserRoleById(Long id);
+    public Optional<UserRole>  getUserRoleById(Integer id);
     public List<UserRole> findAll();
 }
